% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coco.R
\name{coco}
\alias{coco}
\title{The sun of two natural exponential}
\usage{
coco(a, b)
}
\arguments{
\item{a}{any number}

\item{b}{any number}
}
\value{
Returns an interger or float
}
\description{
The sun of two natural exponential
}
\examples{
coco(1,2)
}
