% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HealthData.R
\docType{data}
\name{HealthData}
\alias{HealthData}
\title{World health data}
\format{
264 x 64 dataframe

\describe{
\item{index}{country Names}
\item{column1}{literacy rates}
}
}
\usage{
HealthData
}
\description{
World health data
}
\keyword{datasets}
