% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba.R
\docType{data}
\name{nba}
\alias{nba}
\title{Information about NBA players}
\format{
250 x 5 dataframe

\describe{
\item{index}{ID}
\item{column1}{name}
\item{column2}{Team number}
\item{column3}{Player Id}
\item{column4}{SEASON}
}
}
\usage{
nba
}
\description{
Information about NBA players
}
\keyword{datasets}
