% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\docType{data}
\name{plates}
\alias{plates}
\title{Shanghai plate price}
\format{
190 x 6 dataframe

\describe{
\item{index}{id}
\item{column1}{data}
\item{column2}{average price}
}
}
\usage{
plates
}
\description{
Shanghai plate price
}
\keyword{datasets}
